requirements.txt
numpy~=1.26.2
pytest~=7.4.3
cupy-cuda12x~=12.3.0
scipy~=1.12.0rc1